+++ Libplanet.Tests/Blockchain/BlockChainTest.cs
@@ -13,7 +13,6 @@ using Libplanet.Crypto;
 using Libplanet.Store;
 using Libplanet.Tests.Common.Action;
 using Libplanet.Tests.Store;
-using Libplanet.Tests.Tx;
 using Libplanet.Tx;
 using Serilog;
 using Xunit;
@@ -70,30 +69,36 @@ namespace Libplanet.Tests.Blockchain
                 Transaction<DumbAction>.Create(
                     0,
                     keys[0],
+                    _blockChain.Genesis.Hash,
                     new DumbAction[] { }),
                 Transaction<DumbAction>.Create(
                     1,
                     keys[0],
+                    _blockChain.Genesis.Hash,
                     new DumbAction[] { }),
 
                 // pending txs1
                 Transaction<DumbAction>.Create(
                     1,
                     keys[1],
+                    _blockChain.Genesis.Hash,
                     new DumbAction[] { }),
                 Transaction<DumbAction>.Create(
                     2,
                     keys[1],
+                    _blockChain.Genesis.Hash,
                     new DumbAction[] { }),
 
                 // pending txs2
                 Transaction<DumbAction>.Create(
                     0,
                     keys[2],
+                    _blockChain.Genesis.Hash,
                     new DumbAction[] { }),
                 Transaction<DumbAction>.Create(
                     2,
                     keys[2],
+                    _blockChain.Genesis.Hash,
                     new DumbAction[] { }),
             };
 
@@ -217,6 +222,33 @@ namespace Libplanet.Tests.Blockchain
         }
 
         [Fact]
+        public void StageTransactionWithDifferentGenesis()
+        {
+            Transaction<DumbAction> tx1 = Transaction<DumbAction>.Create(
+                0,
+                new PrivateKey(),
+                _blockChain.Genesis.Hash,
+                new List<DumbAction>());
+            Transaction<DumbAction> tx2 = Transaction<DumbAction>.Create(
+                0,
+                new PrivateKey(),
+                null,
+                new List<DumbAction>());
+            Transaction<DumbAction> tx3 = Transaction<DumbAction>.Create(
+                0,
+                new PrivateKey(),
+                default(HashDigest<SHA256>),
+                new List<DumbAction>());
+
+            _blockChain.StageTransaction(tx1);
+            Assert.Equal(1, _blockChain.GetStagedTransactionIds().Count);
+            Assert.Throws<InvalidTxGenesisHashException>(() => _blockChain.StageTransaction(tx2));
+            Assert.Equal(1, _blockChain.GetStagedTransactionIds().Count);
+            Assert.Throws<InvalidTxGenesisHashException>(() => _blockChain.StageTransaction(tx3));
+            Assert.Equal(1, _blockChain.GetStagedTransactionIds().Count);
+        }
+
+        [Fact]
         public void UnstageTransaction()
         {
             Transaction<DumbAction>[] txs = new[]
@@ -242,6 +274,14 @@ namespace Libplanet.Tests.Blockchain
         [Fact]
         public async void ProcessActions()
         {
+            var genesisBlock = BlockChain<PolymorphicAction<BaseAction>>.MakeGenesisBlock();
+            var store = new DefaultStore(path: null);
+            var chain = new BlockChain<PolymorphicAction<BaseAction>>(
+                new BlockPolicy<PolymorphicAction<BaseAction>>(),
+                store,
+                genesisBlock
+            );
+
             var actions1 = new List<PolymorphicAction<BaseAction>>
             {
                 new Attack
@@ -266,16 +306,10 @@ namespace Libplanet.Tests.Blockchain
             var tx1 = Transaction<PolymorphicAction<BaseAction>>.Create(
                 0,
                 new PrivateKey(),
+                genesisBlock.Hash,
                 actions1
             );
 
-            var genesisBlock = BlockChain<PolymorphicAction<BaseAction>>.MakeGenesisBlock();
-            var store = new DefaultStore(path: null);
-            var chain = new BlockChain<PolymorphicAction<BaseAction>>(
-                new BlockPolicy<PolymorphicAction<BaseAction>>(),
-                store,
-                genesisBlock
-            );
             chain.StageTransaction(tx1);
             await chain.MineBlock(_fx.Address1);
 
@@ -300,6 +334,7 @@ namespace Libplanet.Tests.Blockchain
             var tx2 = Transaction<PolymorphicAction<BaseAction>>.Create(
                 0,
                 new PrivateKey(),
+                genesisBlock.Hash,
                 actions2
             );
 
@@ -313,6 +348,7 @@ namespace Libplanet.Tests.Blockchain
             var tx3 = Transaction<PolymorphicAction<BaseAction>>.Create(
                 0,
                 new PrivateKey(),
+                genesisBlock.Hash,
                 new List<PolymorphicAction<BaseAction>>
                 {
                     new Attack
@@ -1275,7 +1311,7 @@ namespace Libplanet.Tests.Blockchain
                 };
                 Transaction<DumbAction>[] txs =
                 {
-                    Transaction<DumbAction>.Create(0, privateKey, actions),
+                    Transaction<DumbAction>.Create(0, privateKey, chain.Genesis.Hash, actions),
                 };
                 b = TestUtils.MineNext(b, txs);
                 chain.Append(b);
@@ -1535,17 +1571,19 @@ namespace Libplanet.Tests.Blockchain
             long blockIndex = 1;
 
             TestEvaluateAction action = new TestEvaluateAction();
-            var tx1 = Transaction<TestEvaluateAction>.Create(
-                0,
-                fromPrivateKey,
-                new[] { action }
-            );
 
             var store = new DefaultStore(null);
             var chain = TestUtils.MakeBlockChain<TestEvaluateAction>(
                 new BlockPolicy<TestEvaluateAction>(),
                 store
             );
+            var tx1 = Transaction<TestEvaluateAction>.Create(
+                0,
+                fromPrivateKey,
+                chain.Genesis.Hash,
+                new[] { action }
+            );
+
             chain.StageTransaction(tx1);
             await chain.MineBlock(_fx.Address1);
 
@@ -1936,6 +1974,7 @@ namespace Libplanet.Tests.Blockchain
                     Transaction<DumbAction> tx = Transaction<DumbAction>.Create(
                         store.GetTxNonce(chainId, signer),
                         privateKey,
+                        chain.Genesis.Hash,
                         new[] { new DumbAction(addresses[j], index.ToString()) }
                     );
                     b = TestUtils.MineNext(
@@ -2107,7 +2146,6 @@ namespace Libplanet.Tests.Blockchain
         {
             var storeFixture = new DefaultStoreFixture();
             var policy = new NullPolicy<DumbAction>();
-            var txFixture = new TxFixture();
 
             var timestamp = new DateTimeOffset(2019, 11, 20, 0, 0, 0, TimeSpan.Zero);
             var addresses = ImmutableList<Address>.Empty