+++ Libplanet.Tests/Net/SwarmTest.cs
@@ -633,6 +633,7 @@ namespace Libplanet.Tests.Net
             Transaction<DumbAction> tx = Transaction<DumbAction>.Create(
                 0,
                 new PrivateKey(),
+                chainB.Genesis.Hash,
                 new DumbAction[0]
             );
             chainB.StageTransaction(tx);
@@ -675,6 +676,7 @@ namespace Libplanet.Tests.Net
             Transaction<DumbAction> tx = Transaction<DumbAction>.Create(
                 0,
                 new PrivateKey(),
+                chainA.Genesis.Hash,
                 new DumbAction[] { }
             );
 
@@ -772,6 +774,7 @@ namespace Libplanet.Tests.Net
             Transaction<DumbAction> tx = Transaction<DumbAction>.Create(
                 0,
                 new PrivateKey(),
+                chainA.Genesis.Hash,
                 new DumbAction[] { }
             );
 
@@ -829,6 +832,7 @@ namespace Libplanet.Tests.Net
             Transaction<DumbAction> tx = Transaction<DumbAction>.Create(
                 0,
                 new PrivateKey(),
+                blockChains[size - 1].Genesis.Hash,
                 new DumbAction[] { }
             );
 
@@ -1346,6 +1350,7 @@ namespace Libplanet.Tests.Net
                     Transaction<Sleep>.Create(
                         0,
                         new PrivateKey(),
+                        miner1.BlockChain.Genesis.Hash,
                         actions: new[] { new Sleep() }
                     )
                 );
@@ -1496,6 +1501,61 @@ namespace Libplanet.Tests.Net
 
                 Assert.Contains(validTx.Id, swarmB.BlockChain.GetStagedTransactionIds());
                 Assert.DoesNotContain(invalidTx.Id, swarmB.BlockChain.GetStagedTransactionIds());
+            }
+            finally
+            {
+                await StopAsync(swarmA);
+                await StopAsync(swarmB);
+
+                swarmA.Dispose();
+                swarmB.Dispose();
+
+                fx1.Dispose();
+                fx2.Dispose();
+            }
+        }
+
+        [Fact(Timeout = Timeout)]
+        public async Task IgnoreTransactionFromDifferentGenesis()
+        {
+            var validKey = new PrivateKey();
+            bool IsSignerValid(Transaction<DumbAction> tx)
+            {
+                var validAddress = validKey.PublicKey.ToAddress();
+                return tx.Signer.Equals(validAddress);
+            }
+
+            var policy = new BlockPolicy<DumbAction>(doesTransactionFollowPolicy: IsSignerValid);
+            var fx1 = new DefaultStoreFixture();
+            var fx2 = new DefaultStoreFixture();
+
+            var swarmA = CreateSwarm(
+                TestUtils.MakeBlockChain(
+                    policy,
+                    fx1.Store,
+                    privateKey: validKey,
+                    timestamp: DateTimeOffset.MinValue));
+            var swarmB = CreateSwarm(
+                TestUtils.MakeBlockChain(
+                    policy,
+                    fx2.Store,
+                    privateKey: validKey,
+                    timestamp: DateTimeOffset.MinValue.AddSeconds(1)));
+
+            try
+            {
+                var tx = swarmA.BlockChain.MakeTransaction(validKey, new DumbAction[] { });
+
+                await StartAsync(swarmA);
+                await StartAsync(swarmB);
+
+                await BootstrapAsync(swarmA, swarmB.AsPeer);
+
+                swarmA.BroadcastTxs(new[] { tx });
+                await swarmB.TxReceived.WaitAsync();
+
+                Assert.Throws<KeyNotFoundException>(() => swarmB.BlockChain.GetTransaction(tx.Id));
+                Assert.DoesNotContain(tx.Id, swarmB.BlockChain.GetStagedTransactionIds());
             }
             finally
             {
@@ -1797,6 +1857,7 @@ namespace Libplanet.Tests.Net
                 Transaction<DumbAction> tx = Transaction<DumbAction>.Create(
                     0,
                     new PrivateKey(),
+                    sender1.BlockChain.Genesis.Hash,
                     new DumbAction[] { }
                 );
                 sender1.BlockChain.StageTransaction(tx);